+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Upgraded
+- Keyple Distributed Network Lib `2.4.0` -> `2.5.0`
 
 ## [2.4.0] - 2024-06-03
 ### Added
@@ -17,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Logging improvement.
 ### Upgraded
+- Keyple Distributed Remote API `3.0.1` -> `3.1.0`
 - Keyple Distributed Network Lib `2.3.1` -> `2.4.0`
 
 ## [2.3.1] - 2024-04-12